frontend: Merge make-node-form.jsx into nodes.jsx
make-node-form.jsx was only used by nodes.jsx and is actually AWS
specific, so we let's merge these files to make it clear that
makeNodeForm() is only for AWS nodes and to avoid unnecessary exports /
imports.
